Moved budgetary resources request to AccountSpending component; request overview data from the top-level container
