+++ src/js/components/agencyV2/accountSpending/AccountSpending.jsx
@@ -3,8 +3,14 @@
  * Created by Lizzie Salita 5/8/20
  */
 
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
+import { useDispatch } from 'react-redux';
+
+import { setBudgetaryResources } from 'redux/actions/agencyV2/agencyV2Actions';
+import { fetchBudgetaryResources } from 'helpers/agencyV2Helper';
+import BaseAgencyBudgetaryResources from 'models/v2/agency/BaseAgencyBudgetaryResources';
+
 import CountTabContainer from 'containers/agencyV2/accountSpending/CountTabContainer';
 import TableContainer from 'containers/agencyV2/accountSpending/TableContainer';
 
@@ -47,6 +53,29 @@ const tabs = [
 const AccountSpending = ({ agencyId, fy }) => {
     const [activeTab, setActiveTab] = useState('budget_function');
     const subHeading = tabs.find((tab) => tab.type === activeTab).subHeading;
+    const [loading, setLoading] = useState(true);
+    const [error, setError] = useState(false);
+    const dispatch = useDispatch();
+
+    useEffect(() => {
+        setLoading(true);
+        // request budgetary resources data for this agency
+        const budgetaryResourcesRequest = fetchBudgetaryResources(agencyId);
+        budgetaryResourcesRequest.promise
+            .then((res) => {
+                // parse the response using our data model
+                setLoading(false);
+                const budgetaryResources = Object.create(BaseAgencyBudgetaryResources);
+                budgetaryResources.populate(res.data);
+                // store the data model object in Redux
+                dispatch(setBudgetaryResources(budgetaryResources));
+            }).catch((err) => {
+                setError(true);
+                setLoading(false);
+                console.error(err);
+            });
+    }, [agencyId]);
+
     return (
         <div className="body__content">
             <div className="count-tabs">
