+++ src/js/containers/router/RouterRoutes.js
@@ -28,7 +28,7 @@ const StateLandingPage = React.lazy(() => import('components/stateLanding/StateL
 const StateContainer = React.lazy(() => import('containers/state/StateContainer').then((comp) => comp));
 const RecipientLandingPage = React.lazy(() => import('components/recipientLanding/RecipientLandingPage').then((comp) => comp));
 const RecipientContainer = React.lazy(() => import('containers/recipient/RecipientContainer').then((comp) => comp));
-const AgencyProfileV2 = React.lazy(() => import('containers/agency/v2/AgencyContainerV2').then((comp) => comp));
+const AgencyProfileV2 = React.lazy(() => import('containers/agencyV2/AgencyContainerV2').then((comp) => comp));
 const Covid19Container = React.lazy(() => import('containers/covid19/Covid19Container').then((comp) => comp));
 const DataSourcesAndMethodologiesPage = React.lazy(() => import('components/covid19/DataSourcesAndMethodologiesPage').then((comp) => comp));
 const AboutTheDataPage = React.lazy(() => import('components/aboutTheData/AboutTheDataPage').then((comp) => comp));
